vertexbuffers

+++ Bindings/include/Bindings/Filament/VertexBuffer.h
@@ -15,7 +15,7 @@
 @property (nonatomic, readonly, nonnull) void* buffer  NS_SWIFT_UNAVAILABLE("Don't access the raw pointers");
 - (nonnull id) init: (nonnull void*) buffer NS_SWIFT_UNAVAILABLE("Create a new renderer with engine.createRenderer");
 - (nonnull id) init NS_UNAVAILABLE;
-- (int) getVertexCount;
+- (size_t) getVertexCount;
 - (void) setBufferAt: (nonnull Engine*) engine :(int) bufferIndex :(nonnull NSData*) data :(int) byteOffset;
 - (void) setBufferAt: (nonnull Engine*) engine :(int) bufferIndex :(nonnull NSData*) data;
